requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
numpy
gensim
plotly
flask
pandas
nbformat>=4.2.0
nltk
streamlit
scikit-learn
wordcloud